requirements.txt
streamlit
pandas
plotly
folium
streamlit-folium
bcrypt
streamlit-oauth
google-generativeai